feat: add display roles on project grid view
